+++ packages/google-analytics/CHANGELOG.md
@@ -4,4 +4,4 @@
 
 ### Patch Changes
 
-- 76fe5ba: Initialize package
+- b9c1dbb: Initialize package