+++ packages/reddit/CHANGELOG.md
@@ -4,6 +4,6 @@
 
 ### Patch Changes
 
-- 76fe5ba: Initialize package
-- Updated dependencies [76fe5ba]
+- b9c1dbb: Initialize package
+- Updated dependencies [b9c1dbb]
   - @microfox/reddit-oauth@1.0.1